github-app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
PyJWT[crypto]>=2.8.0
pydantic-settings>=2.1.0
